refactor: Move function.json logic into new FunctionBuildService
