+++ src/models/functionApp.ts
@@ -1,4 +1,8 @@
-export interface FunctionAppHttpTriggerConfig {
+export interface FunctionBindingConfig {
+  type?: string;
+}
+
+export interface FunctionHttpBindingConfig extends FunctionBindingConfig {
   authLevel: string;
   methods: string[];
   route: string;
